Added signal handling

+++ Functions.h
@@ -850,7 +850,7 @@ comp->mov  (comp->r_rax, (int32_t)(void*)&VirtualProcessor_t::Jmp);
 comp->mov  (comp->r_rcx, (int32_t)(void*)this);
 comp->call (comp->r_rax);
 comp->jmp (0);
-JmpPatchRequestLine (comp, comp->Size() - 4, arg.arg1);
+JmpPatchRequestLine (comp, (int64_t)(comp->Size() - 4), arg.arg1);
 FUNCTION_END
 
 FUNCTION_BEGIN (Jne, 1, 0, ARG_LABEL)
@@ -859,7 +859,7 @@ comp->mov  (comp->r_rcx, (int32_t)(void*)this);
 comp->call (comp->r_rax);
 comp->cmp (&instance_->cmpr_flag_, '\0');
 comp->jne (0);
-JmpPatchRequestLine (comp, comp->Size() - 4, arg.arg1);
+JmpPatchRequestLine (comp, (int64_t)(comp->Size() - 4), arg.arg1);
 comp->inc  (&instance_->run_line_);
 FUNCTION_END
 
@@ -869,7 +869,7 @@ comp->mov  (comp->r_rcx, (int32_t)(void*)this);
 comp->call (comp->r_rax);
 comp->cmp (&instance_->cmpr_flag_, '\0');
 comp->je (0);
-JmpPatchRequestLine (comp, comp->Size() - 4, arg.arg1);
+JmpPatchRequestLine (comp, (int64_t)(comp->Size() - 4), arg.arg1);
 comp->inc  (&instance_->run_line_);
 FUNCTION_END
 
@@ -879,7 +879,7 @@ comp->mov  (comp->r_rcx, (int32_t)(void*)this);
 comp->call (comp->r_rax);
 comp->cmp (&instance_->cmpr_flag_, '\0');
 comp->jg (0);
-JmpPatchRequestLine (comp, comp->Size() - 4, arg.arg1);
+JmpPatchRequestLine (comp, (int64_t)(comp->Size() - 4), arg.arg1);
 comp->inc  (&instance_->run_line_);
 FUNCTION_END
 
@@ -889,7 +889,7 @@ comp->mov  (comp->r_rcx, (int32_t)(void*)this);
 comp->call (comp->r_rax);
 comp->cmp (&instance_->cmpr_flag_, '\0');
 comp->jge (0);
-JmpPatchRequestLine (comp, comp->Size() - 4, arg.arg1);
+JmpPatchRequestLine (comp, (int64_t)(comp->Size() - 4), arg.arg1);
 comp->inc  (&instance_->run_line_);
 FUNCTION_END
 
@@ -899,7 +899,7 @@ comp->mov  (comp->r_rcx, (int32_t)(void*)this);
 comp->call (comp->r_rax);
 comp->cmp (&instance_->cmpr_flag_, '\0');
 comp->jl (0);
-JmpPatchRequestLine (comp, comp->Size() - 4, arg.arg1);
+JmpPatchRequestLine (comp, (int64_t)(comp->Size() - 4), arg.arg1);
 comp->inc  (&instance_->run_line_);
 FUNCTION_END
 
@@ -909,7 +909,7 @@ comp->mov  (comp->r_rcx, (int32_t)(void*)this);
 comp->call (comp->r_rax);
 comp->cmp (&instance_->cmpr_flag_, '\0');
 comp->jle (0);
-JmpPatchRequestLine (comp, comp->Size() - 4, arg.arg1);
+JmpPatchRequestLine (comp, (int64_t)(comp->Size() - 4), arg.arg1);
 comp->inc  (&instance_->run_line_);
 FUNCTION_END
 
@@ -926,7 +926,7 @@ comp->mov   (comp->r_rcx, (int32_t)(void*)this);
 comp->call  (comp->r_rax);
 comp->inc  (&instance_->run_line_);
 comp->callr (0);
-JmpPatchRequestLine (comp, comp->Size() - 4, arg.arg1 + 1);
+JmpPatchRequestLine (comp, (int64_t)(comp->Size() - 4), arg.arg1 + 1);
 FUNCTION_END
 
 FUNCTION_BEGIN (Decr, 3, 0, ARG_VAR _ ARG_VAR_MEMBER _ ARG_REG)
